Rename astrometry v2 API and drop astrometry2
